refactored to sdk
ci fixes
