+++ Cargo.toml
@@ -136,7 +136,7 @@ members = [
     "crates/trie/trie",
     "examples/beacon-api-sidecar-fetcher/",
     "examples/beacon-api-sse/",
-    "examples/bsc-p2p",
+    "examples/bsc-sdk",
     "examples/custom-dev-node/",
     "examples/custom-node/",
     "examples/custom-engine-types/",
