--- examples/bsc-p2p/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod block_import;
-pub mod chainspec;
-pub mod evm;
-pub mod handshake;
-pub mod upgrade_status;
