+++ examples/bsc-sdk/src/evm/mod.rs
@@ -2,5 +2,5 @@ mod api;
 mod config;
 mod handler;
 mod precompiles;
-mod spec;
+pub mod spec;
 mod transaction;
