+++ examples/bsc-sdk/src/node/network/handshake.rs
@@ -1,4 +1,4 @@
-use crate::upgrade_status::{UpgradeStatus, UpgradeStatusExtension};
+use super::upgrade_status::{UpgradeStatus, UpgradeStatusExtension};
 use alloy_rlp::Decodable;
 use futures::SinkExt;
 use reth_eth_wire::{
