+++ examples/bsc-sdk/tests/it/p2p.rs
@@ -1,6 +1,6 @@
-use example_bsc_p2p::{
-    chainspec::{boot_nodes, bsc_chain_spec, head},
-    handshake::BscHandshake,
+use example_bsc_sdk::{
+    chainspec::{bsc_chain_spec, head},
+    node::network::{boot_nodes, handshake::BscHandshake},
 };
 use reth_chainspec::NamedChain;
 use reth_discv4::Discv4ConfigBuilder;
